refactor(ledger): collocate the asset route registrars :hammer:
The asset resource had its two registrars in two different files: the Huma
terminal registrar in asset_handler.go and the Fiber guard-chain registrar in
routes.go. Both halves now live in asset_routes.go, so the whole asset route
surface is one grep target.

asset_handler.go keeps the handler struct, the request/response envelopes and
the transport-agnostic cores. routes.go keeps only the shared primitives
(protectedMidaz, the routeXxx helpers, registerRoute).

The new file needs stdlib net/http for the Huma operation constants and
pkg/net/http for ProtectedRouteOptions, so the midaz import takes the pkgHTTP
alias every handler file already uses.

No route, authz tuple, operation ID or path changed.
