+++ components/ledger/internal/adapters/http/in/asset_handler.go
@@ -330,88 +330,3 @@ func (handler *AssetHandler) CountAssets(ctx context.Context, in *CountAssetsReq
 		ContentLength: "0",
 	}, nil
 }
-
-// RegisterAssetRoutes registers the six asset operations on the shared
-// Huma API. It is the per-file seam registerAssetRoutesToApp calls; the auth + tenant +
-// ParseUUIDPathParameters middleware chain for these routes is attached in
-// registerAssetRoutesToApp (Fiber-level) BEFORE the Huma terminal, not here.
-//
-// Paths are GROUP-RELATIVE: the Huma API is bound to a versioned Fiber group, so the
-// humafiber adapter registers on that group and Fiber prepends the version prefix.
-//
-// opSuffix distinguishes the operation IDs one version group publishes from another's —
-// see routeOpSuffixV1. A straight v1/v2 mirror reuses the same handler methods and the
-// same input/output types, so only the operation IDs differ between the twins.
-func RegisterAssetRoutes(api huma.API, h *AssetHandler, opSuffix string) {
-	const (
-		listPath  = "/organizations/{organization_id}/ledgers/{ledger_id}/assets"
-		idPath    = listPath + "/{id}"
-		countPath = listPath + "/metrics/count"
-		tag       = "Assets"
-	)
-
-	huma.Register(api, huma.Operation{
-		OperationID: "createAsset" + opSuffix,
-		Method:      http.MethodPost,
-		Path:        listPath,
-		Summary:     "Create a new asset",
-		Tags:        []string{tag},
-		Security:    secAssetBearer,
-		// Body validated imperatively (http.DecodeAndValidate) — see file header.
-		SkipValidateBody: true,
-		DefaultStatus:    http.StatusCreated,
-	}, h.CreateAsset)
-	attachTypedRequestBody[mmodel.CreateAssetInput](api, "createAsset"+opSuffix)
-
-	huma.Register(api, huma.Operation{
-		OperationID: "listAssets" + opSuffix,
-		Method:      http.MethodGet,
-		Path:        listPath,
-		Summary:     "List all assets",
-		Tags:        []string{tag},
-		Security:    secAssetBearer,
-	}, h.ListAssets)
-
-	huma.Register(api, huma.Operation{
-		OperationID: "getAssetByID" + opSuffix,
-		Method:      http.MethodGet,
-		Path:        idPath,
-		Summary:     "Retrieve a specific asset",
-		Tags:        []string{tag},
-		Security:    secAssetBearer,
-	}, h.GetAssetByID)
-
-	huma.Register(api, huma.Operation{
-		OperationID:      "updateAsset" + opSuffix,
-		Method:           http.MethodPatch,
-		Path:             idPath,
-		Summary:          "Update an asset",
-		Tags:             []string{tag},
-		Security:         secAssetBearer,
-		SkipValidateBody: true, // body validated imperatively — see createAsset.
-	}, h.UpdateAsset)
-	attachTypedRequestBody[mmodel.UpdateAssetInput](api, "updateAsset"+opSuffix)
-
-	huma.Register(api, huma.Operation{
-		OperationID: "deleteAsset" + opSuffix,
-		Method:      http.MethodDelete,
-		Path:        idPath,
-		Summary:     "Delete an asset",
-		Tags:        []string{tag},
-		Security:    secAssetBearer,
-		// DefaultStatus 204 + an Out struct with no Body field => bodiless 204.
-		DefaultStatus: http.StatusNoContent,
-	}, h.DeleteAssetByID)
-
-	huma.Register(api, huma.Operation{
-		OperationID: "countAssets" + opSuffix,
-		Method:      http.MethodHead,
-		Path:        countPath,
-		Summary:     "Count total assets",
-		Tags:        []string{tag},
-		Security:    secAssetBearer,
-		// HEAD count: X-Total-Count header + empty 204 body (Content-Length 0 set
-		// on the Out struct), matching the Fiber http.NoContent + header path.
-		DefaultStatus: http.StatusNoContent,
-	}, h.CountAssets)
-}
